requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
colorlog==4.8.0
filetype==1.0.7
har2case==0.3.1
httprunner==2.5.7
idna==3.2
Jinja2==2.11.3
jsonpath==0.82
jsonschema==3.2.0
MarkupSafe==2.0.1
pyrsistent==0.18.0
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
sentry-sdk==0.13.5
six==1.16.0
urllib3==1.26.6
